pyproject.toml
[project]
name = "wagtailcodeblock"
authors = [
    {name = "John Higgins", email = "johnmatthiggins@gmail.com"},
    {name = "Tim Allen", email = "tallen@wharton.upenn.edu"},
]
description = "Wagtail Code Block provides PrismJS syntax highlighting in Wagtail."
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.7"
keywords = ["wagtail", "cms", "contact", "syntax", "code", "highlighting", "highlighter"]
license = {text = "BSD-3-Clause"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django",
    "Framework :: Wagtail",
    "Framework :: Wagtail :: 3",
    "Framework :: Wagtail :: 4",
    "Framework :: Wagtail :: 5",
    "Framework :: Wagtail :: 6",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "wagtail>=4",
]

[project.optional-dependencies]
dev = [
  "django-coverage-plugin",
  "ipython",
  "ruff",
  "pytest-coverage",
  "pytest-django",
]

[project.urls]
"Homepage" = "https://github.com/FlipperPA/wagtailcodeblock"
"Repository" = "https://github.com/FlipperPA/wagtailcodeblock"
"Documentation" = "https://github.com/FlipperPA/wagtailcodeblock"

[build-system]
requires = ["setuptools>=67", "setuptools_scm>=7", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "wagtailcodeblock/_version.py"

[tool.pytest.ini_options]
addopts = "--cov --cov-report=html"
python_files = "tests.py test_*.py"
DJANGO_SETTINGS_MODULE = "tests.settings"

[tool.coverage.run]
plugins = ["django_coverage_plugin"]
include = ["wagtailcodeblock/*"]
